+++ lib/event/reactor.c
@@ -70,6 +70,7 @@ static struct spdk_scheduler *g_scheduler;
 static struct spdk_scheduler *g_new_scheduler;
 static struct spdk_reactor *g_scheduling_reactor;
 static uint64_t g_scheduler_period;
+static uint32_t g_scheduler_core_number;
 static struct spdk_scheduler_core_info *g_core_infos = NULL;
 
 TAILQ_HEAD(, spdk_governor) g_governor_list
@@ -153,6 +154,12 @@ _spdk_scheduler_period_set(uint64_t period)
 }
 
 void
+_spdk_scheduler_disable(void)
+{
+	g_scheduler_period = 0;
+}
+
+void
 _spdk_scheduler_list_add(struct spdk_scheduler *scheduler)
 {
 	if (_scheduler_find(scheduler->name)) {
@@ -172,6 +179,7 @@ reactor_construct(struct spdk_reactor *reactor, uint32_t lcore)
 
 	TAILQ_INIT(&reactor->threads);
 	reactor->thread_count = 0;
+	spdk_cpuset_zero(&reactor->notify_cpuset);
 
 	reactor->events = spdk_ring_create(SPDK_RING_TYPE_MP_SC, 65536, SPDK_ENV_SOCKET_ID_ANY);
 	if (reactor->events == NULL) {
@@ -179,8 +187,26 @@ reactor_construct(struct spdk_reactor *reactor, uint32_t lcore)
 		assert(false);
 	}
 
+	/* Always initialize interrupt facilities for reactor */
+	if (reactor_interrupt_init(reactor) != 0) {
+		/* Reactor interrupt facilities are necessary if seting app to interrupt mode. */
+		if (spdk_interrupt_mode_is_enabled()) {
+			SPDK_ERRLOG("Failed to prepare intr facilities\n");
+			assert(false);
+		}
+		return;
+	}
+
+	/* If application runs with full interrupt ability,
+	 * all reactors are going to run in interrupt mode.
+	 */
 	if (spdk_interrupt_mode_is_enabled()) {
-		reactor_interrupt_init(reactor);
+		uint32_t i;
+
+		SPDK_ENV_FOREACH_CORE(i) {
+			spdk_cpuset_set_cpu(&reactor->notify_cpuset, i, true);
+		}
+		reactor->in_interrupt = true;
 	}
 }
 
@@ -205,6 +231,12 @@ spdk_reactor_get(uint32_t lcore)
 	}
 
 	return reactor;
+}
+
+struct spdk_reactor *
+_spdk_get_scheduling_reactor(void)
+{
+	return g_scheduling_reactor;
 }
 
 static int reactor_thread_op(struct spdk_thread *thread, enum spdk_thread_op op);
@@ -298,9 +330,7 @@ spdk_reactors_fini(void)
 			spdk_ring_free(reactor->events);
 		}
 
-		if (reactor->interrupt_mode) {
-			reactor_interrupt_fini(reactor);
-		}
+		reactor_interrupt_fini(reactor);
 
 		if (g_core_infos != NULL) {
 			free(g_core_infos[i].threads);
@@ -313,6 +343,124 @@ spdk_reactors_fini(void)
 	g_reactors = NULL;
 	free(g_core_infos);
 	g_core_infos = NULL;
+}
+
+static void _reactor_set_interrupt_mode(void *arg1, void *arg2);
+
+static void
+_reactor_set_notify_cpuset(void *arg1, void *arg2)
+{
+	struct spdk_reactor *target = arg1;
+	struct spdk_reactor *reactor = spdk_reactor_get(spdk_env_get_current_core());
+
+	spdk_cpuset_set_cpu(&reactor->notify_cpuset, target->lcore, target->new_in_interrupt);
+}
+
+static void
+_reactor_set_notify_cpuset_cpl(void *arg1, void *arg2)
+{
+	struct spdk_reactor *target = arg1;
+
+	if (target->new_in_interrupt == false) {
+		target->set_interrupt_mode_in_progress = false;
+		spdk_thread_send_msg(_spdk_get_app_thread(), target->set_interrupt_mode_cb_fn,
+				     target->set_interrupt_mode_cb_arg);
+	} else {
+		struct spdk_event *ev;
+
+		ev = spdk_event_allocate(target->lcore, _reactor_set_interrupt_mode, target, NULL);
+		assert(ev);
+		spdk_event_call(ev);
+	}
+}
+
+static void
+_reactor_set_interrupt_mode(void *arg1, void *arg2)
+{
+	struct spdk_reactor *target = arg1;
+
+	assert(target == spdk_reactor_get(spdk_env_get_current_core()));
+	assert(target != NULL);
+	assert(target->in_interrupt != target->new_in_interrupt);
+	assert(TAILQ_EMPTY(&target->threads));
+	SPDK_DEBUGLOG(reactor, "Do reactor set on core %u from %s to state %s\n",
+		      target->lcore, !target->in_interrupt ? "intr" : "poll", target->new_in_interrupt ? "intr" : "poll");
+
+	target->in_interrupt = target->new_in_interrupt;
+
+	if (target->new_in_interrupt == false) {
+		spdk_for_each_reactor(_reactor_set_notify_cpuset, target, NULL, _reactor_set_notify_cpuset_cpl);
+	} else {
+		uint64_t notify = 1;
+		int rc = 0;
+
+		/* Always trigger spdk_event and resched event in case of race condition */
+		rc = write(target->events_fd, &notify, sizeof(notify));
+		if (rc < 0) {
+			SPDK_ERRLOG("failed to notify event queue: %s.\n", spdk_strerror(errno));
+		}
+		rc = write(target->resched_fd, &notify, sizeof(notify));
+		if (rc < 0) {
+			SPDK_ERRLOG("failed to notify reschedule: %s.\n", spdk_strerror(errno));
+		}
+
+		target->set_interrupt_mode_in_progress = false;
+		spdk_thread_send_msg(_spdk_get_app_thread(), target->set_interrupt_mode_cb_fn,
+				     target->set_interrupt_mode_cb_arg);
+	}
+}
+
+int
+spdk_reactor_set_interrupt_mode(uint32_t lcore, bool new_in_interrupt,
+				spdk_reactor_set_interrupt_mode_cb cb_fn, void *cb_arg)
+{
+	struct spdk_reactor *target;
+
+	target = spdk_reactor_get(lcore);
+	if (target == NULL) {
+		return -EINVAL;
+	}
+
+	if (spdk_get_thread() != _spdk_get_app_thread()) {
+		SPDK_ERRLOG("It is only permitted within spdk application thread.\n");
+		return -EPERM;
+	}
+
+	if (target->in_interrupt == new_in_interrupt) {
+		return 0;
+	}
+
+	if (target->set_interrupt_mode_in_progress) {
+		SPDK_NOTICELOG("Reactor(%u) is already in progress to set interrupt mode\n", lcore);
+		return -EBUSY;
+	}
+	target->set_interrupt_mode_in_progress = true;
+
+	target->new_in_interrupt = new_in_interrupt;
+	target->set_interrupt_mode_cb_fn = cb_fn;
+	target->set_interrupt_mode_cb_arg = cb_arg;
+
+	SPDK_DEBUGLOG(reactor, "Starting reactor event from %d to %d\n",
+		      spdk_env_get_current_core(), lcore);
+
+	if (new_in_interrupt == false) {
+		/* For potential race cases, when setting the reactor to poll mode,
+		 * first change the mode of the reactor and then clear the corresponding
+		 * bit of the notify_cpuset of each reactor.
+		 */
+		struct spdk_event *ev;
+
+		ev = spdk_event_allocate(lcore, _reactor_set_interrupt_mode, target, NULL);
+		assert(ev);
+		spdk_event_call(ev);
+	} else {
+		/* For race caces, when setting the reactor to interrupt mode, first set the
+		 * corresponding bit of the notify_cpuset of each reactor and then change the mode.
+		 */
+		spdk_for_each_reactor(_reactor_set_notify_cpuset, target, NULL, _reactor_set_notify_cpuset_cpl);
+	}
+
+	return 0;
 }
 
 struct spdk_event *
@@ -345,6 +493,8 @@ spdk_event_call(struct spdk_event *event)
 {
 	int rc;
 	struct spdk_reactor *reactor;
+	struct spdk_reactor *local_reactor = NULL;
+	uint32_t current_core = spdk_env_get_current_core();
 
 	reactor = spdk_reactor_get(event->lcore);
 
@@ -356,7 +506,16 @@ spdk_event_call(struct spdk_event *event)
 		assert(false);
 	}
 
-	if (reactor->interrupt_mode) {
+	if (current_core != SPDK_ENV_LCORE_ID_ANY) {
+		local_reactor = spdk_reactor_get(current_core);
+	}
+
+	/* If spdk_event_call isn't called on a reactor, always send a notification.
+	 * If it is called on a reactor, send a notification if the destination reactor
+	 * is indicated in interrupt mode state.
+	 */
+	if (spdk_unlikely(local_reactor == NULL) ||
+	    spdk_unlikely(spdk_cpuset_get_cpu(&local_reactor->notify_cpuset, event->lcore))) {
 		uint64_t notify = 1;
 
 		rc = write(reactor->events_fd, &notify, sizeof(notify));
@@ -383,7 +542,8 @@ event_queue_run_batch(struct spdk_reactor *reactor)
 	memset(events, 0, sizeof(events));
 #endif
 
-	if (reactor->interrupt_mode) {
+	/* Operate event notification if this reactor currently runs in interrupt state */
+	if (spdk_unlikely(reactor->in_interrupt)) {
 		uint64_t notify = 1;
 		int rc;
 
@@ -522,22 +682,54 @@ _threads_reschedule(struct spdk_scheduler_core_info *cores_info)
 }
 
 static void
-_reactors_scheduler_fini(void *arg1, void *arg2)
+_reactors_scheduler_fini(void)
 {
 	struct spdk_reactor *reactor;
 	uint32_t i;
 
+	/* Reschedule based on the balancing output */
+	_threads_reschedule(g_core_infos);
+
+	SPDK_ENV_FOREACH_CORE(i) {
+		reactor = spdk_reactor_get(i);
+		assert(reactor != NULL);
+		reactor->flags.is_scheduling = false;
+	}
+}
+
+static void
+_reactors_scheduler_update_core_mode(void *ctx)
+{
+	struct spdk_reactor *reactor;
+	int rc = 0;
+
+	g_scheduler_core_number = spdk_env_get_next_core(g_scheduler_core_number);
+	if (g_scheduler_core_number == SPDK_ENV_LCORE_ID_ANY) {
+		_reactors_scheduler_fini();
+		return;
+	}
+
+	reactor = spdk_reactor_get(g_scheduler_core_number);
+	if (reactor->in_interrupt != g_core_infos[g_scheduler_core_number].interrupt_mode) {
+		/* Switch next found reactor to new state */
+		rc = spdk_reactor_set_interrupt_mode(g_scheduler_core_number,
+						     g_core_infos[g_scheduler_core_number].interrupt_mode, _reactors_scheduler_update_core_mode, NULL);
+		if (rc == 0) {
+			return;
+		}
+	}
+
+	_reactors_scheduler_update_core_mode(NULL);
+}
+
+static void
+_reactors_scheduler_balance(void *arg1, void *arg2)
+{
 	if (g_reactor_state == SPDK_REACTOR_STATE_RUNNING) {
 		g_scheduler->balance(g_core_infos, g_reactor_count, &g_governor);
 
-		/* Reschedule based on the balancing output */
-		_threads_reschedule(g_core_infos);
-
-		SPDK_ENV_FOREACH_CORE(i) {
-			reactor = spdk_reactor_get(i);
-			assert(reactor != NULL);
-			reactor->flags.is_scheduling = false;
-		}
+		g_scheduler_core_number = spdk_env_get_first_core();
+		_reactors_scheduler_update_core_mode(NULL);
 	}
 }
 
@@ -572,6 +764,7 @@ _reactors_scheduler_gather_metrics(void *arg1, void *arg2)
 	core_info->lcore = reactor->lcore;
 	core_info->core_idle_tsc = reactor->idle_tsc;
 	core_info->core_busy_tsc = reactor->busy_tsc;
+	core_info->interrupt_mode = reactor->in_interrupt;
 
 	SPDK_DEBUGLOG(reactor, "Gathering metrics on %u\n", reactor->lcore);
 
@@ -614,7 +807,7 @@ _reactors_scheduler_gather_metrics(void *arg1, void *arg2)
 	/* If we've looped back around to the scheduler thread, move to the next phase */
 	if (next_core == g_scheduling_reactor->lcore) {
 		/* Phase 2 of scheduling is rebalancing - deciding which threads to move where */
-		evt = spdk_event_allocate(next_core, _reactors_scheduler_fini, NULL, NULL);
+		evt = spdk_event_allocate(next_core, _reactors_scheduler_balance, NULL, NULL);
 		spdk_event_call(evt);
 		return;
 	}
@@ -638,7 +831,8 @@ reactor_post_process_lw_thread(struct spdk_reactor *reactor, struct spdk_lw_thre
 		assert(reactor->thread_count > 0);
 		reactor->thread_count--;
 
-		if (reactor->interrupt_mode) {
+		/* Operate thread intr if running with full interrupt ability */
+		if (spdk_interrupt_mode_is_enabled()) {
 			efd = spdk_thread_get_interrupt_fd(thread);
 			spdk_fd_group_remove(reactor->fgrp, efd);
 		}
@@ -653,7 +847,8 @@ reactor_post_process_lw_thread(struct spdk_reactor *reactor, struct spdk_lw_thre
 			assert(reactor->thread_count > 0);
 			reactor->thread_count--;
 
-			if (reactor->interrupt_mode) {
+			/* Operate thread intr if running with full interrupt ability */
+			if (spdk_interrupt_mode_is_enabled()) {
 				efd = spdk_thread_get_interrupt_fd(thread);
 				spdk_fd_group_remove(reactor->fgrp, efd);
 			}
@@ -728,13 +923,15 @@ reactor_run(void *arg)
 	reactor->tsc_last = spdk_get_ticks();
 
 	while (1) {
-		if (spdk_unlikely(reactor->interrupt_mode)) {
+		/* Execute interrupt process fn if this reactor currently runs in interrupt state */
+		if (spdk_unlikely(reactor->in_interrupt)) {
 			reactor_interrupt_run(reactor);
 		} else {
 			_reactor_run(reactor);
 		}
 
-		if (spdk_unlikely((reactor->tsc_last - last_sched) > g_scheduler_period &&
+		if (spdk_unlikely(g_scheduler_period > 0 &&
+				  (reactor->tsc_last - last_sched) > g_scheduler_period &&
 				  reactor == g_scheduling_reactor &&
 				  !reactor->flags.is_scheduling)) {
 			if (spdk_unlikely(g_scheduler != g_new_scheduler)) {
@@ -769,7 +966,9 @@ reactor_run(void *arg)
 				TAILQ_REMOVE(&reactor->threads, lw_thread, link);
 				assert(reactor->thread_count > 0);
 				reactor->thread_count--;
-				if (reactor->interrupt_mode) {
+
+				/* Operate thread intr if running with full interrupt ability */
+				if (spdk_interrupt_mode_is_enabled()) {
 					int efd = spdk_thread_get_interrupt_fd(thread);
 
 					spdk_fd_group_remove(reactor->fgrp, efd);
@@ -851,12 +1050,18 @@ spdk_reactors_stop(void *arg1)
 	uint32_t i;
 	int rc;
 	struct spdk_reactor *reactor;
+	struct spdk_reactor *local_reactor;
 	uint64_t notify = 1;
 
 	g_reactor_state = SPDK_REACTOR_STATE_EXITING;
+	local_reactor = spdk_reactor_get(spdk_env_get_current_core());
 
-	if (spdk_interrupt_mode_is_enabled()) {
-		SPDK_ENV_FOREACH_CORE(i) {
+	SPDK_ENV_FOREACH_CORE(i) {
+		/* If spdk_event_call isn't called  on a reactor, always send a notification.
+		 * If it is called on a reactor, send a notification if the destination reactor
+		 * is indicated in interrupt mode state.
+		 */
+		if (local_reactor == NULL || spdk_cpuset_get_cpu(&local_reactor->notify_cpuset, i)) {
 			reactor = spdk_reactor_get(i);
 			assert(reactor != NULL);
 			rc = write(reactor->events_fd, &notify, sizeof(notify));
@@ -888,14 +1093,14 @@ _schedule_thread(void *arg1, void *arg2)
 	int efd;
 
 	current_core = spdk_env_get_current_core();
-
 	reactor = spdk_reactor_get(current_core);
 	assert(reactor != NULL);
 
 	TAILQ_INSERT_TAIL(&reactor->threads, lw_thread, link);
 	reactor->thread_count++;
 
-	if (reactor->interrupt_mode) {
+	/* Operate thread intr if running with full interrupt ability */
+	if (spdk_interrupt_mode_is_enabled()) {
 		int rc;
 		struct spdk_thread *thread;
 
@@ -917,6 +1122,10 @@ _reactor_schedule_thread(struct spdk_thread *thread)
 	struct spdk_event *evt = NULL;
 	struct spdk_cpuset *cpumask;
 	uint32_t i;
+	struct spdk_reactor *local_reactor = NULL;
+	uint32_t current_lcore = spdk_env_get_current_core();
+	struct spdk_cpuset polling_cpumask;
+	struct spdk_cpuset valid_cpumask;
 
 	cpumask = spdk_thread_get_cpumask(thread);
 
@@ -926,6 +1135,46 @@ _reactor_schedule_thread(struct spdk_thread *thread)
 	last_stats = lw_thread->last_stats;
 	memset(lw_thread, 0, sizeof(*lw_thread));
 	lw_thread->last_stats = last_stats;
+
+	if (current_lcore != SPDK_ENV_LCORE_ID_ANY) {
+		local_reactor = spdk_reactor_get(current_lcore);
+		assert(local_reactor);
+	}
+
+	/* When interrupt ability of spdk_thread is not enabled and the current
+	 * reactor runs on DPDK thread, skip reactors which are in interrupt mode.
+	 */
+	if (!spdk_interrupt_mode_is_enabled() && local_reactor != NULL) {
+		/* Get the cpumask of all reactors in polling */
+		spdk_cpuset_zero(&polling_cpumask);
+		SPDK_ENV_FOREACH_CORE(i) {
+			spdk_cpuset_set_cpu(&polling_cpumask, i, true);
+		}
+		spdk_cpuset_xor(&polling_cpumask, &local_reactor->notify_cpuset);
+
+		if (core == SPDK_ENV_LCORE_ID_ANY) {
+			/* Get the cpumask of all valid reactors which are suggested and also in polling */
+			spdk_cpuset_copy(&valid_cpumask, &polling_cpumask);
+			spdk_cpuset_and(&valid_cpumask, spdk_thread_get_cpumask(thread));
+
+			/* If there are any valid reactors, spdk_thread should be scheduled
+			 * into one of the valid reactors.
+			 * If there is no valid reactors, spdk_thread should be scheduled
+			 * into one of the polling reactors.
+			 */
+			if (spdk_cpuset_count(&valid_cpumask) != 0) {
+				cpumask = &valid_cpumask;
+			} else {
+				cpumask = &polling_cpumask;
+			}
+		} else if (!spdk_cpuset_get_cpu(&polling_cpumask, core)) {
+			/* If specified reactor is not in polling, spdk_thread should be scheduled
+			 * into one of the polling reactors.
+			 */
+			core = SPDK_ENV_LCORE_ID_ANY;
+			cpumask = &polling_cpumask;
+		}
+	}
 
 	pthread_mutex_lock(&g_scheduler_mtx);
 	if (core == SPDK_ENV_LCORE_ID_ANY) {
@@ -975,7 +1224,9 @@ _reactor_request_thread_reschedule(struct spdk_thread *thread)
 	current_core = spdk_env_get_current_core();
 	reactor = spdk_reactor_get(current_core);
 	assert(reactor != NULL);
-	if (reactor->interrupt_mode) {
+
+	/* Send a notification if the destination reactor is indicated in intr mode state */
+	if (spdk_unlikely(spdk_cpuset_get_cpu(&reactor->notify_cpuset, reactor->lcore))) {
 		uint64_t notify = 1;
 
 		if (write(reactor->resched_fd, &notify, sizeof(notify)) < 0) {
@@ -1086,7 +1337,7 @@ reactor_schedule_thread_event(void *arg)
 	uint32_t count = 0;
 	uint64_t notify = 1;
 
-	assert(reactor->interrupt_mode);
+	assert(reactor->in_interrupt);
 
 	if (read(reactor->resched_fd, &notify, sizeof(notify)) < 0) {
 		SPDK_ERRLOG("failed to acknowledge reschedule: %s.\n", spdk_strerror(errno));
@@ -1141,11 +1392,11 @@ reactor_interrupt_init(struct spdk_reactor *reactor)
 		goto err;
 	}
 
-	reactor->interrupt_mode = true;
 	return 0;
 
 err:
 	spdk_fd_group_destroy(reactor->fgrp);
+	reactor->fgrp = NULL;
 	return rc;
 }
 #else