+++ lib/sock/sock.c
@@ -93,7 +93,10 @@ sock_map_insert(int placement_id, struct spdk_sock_group *group)
 }
 
 /* Release a reference to the group for a given placement_id.
- * If the reference count is 0, remove the group.
+ * We use lazy free method. If a placement_id with a sock is associated with the group,
+ * it will possibly be associated again by another sock with the same placement_id. And
+ * there will no memory leak, because if a polling group is destroyed, the
+ * sock_remove_sock_group_from_map_table will be called to guarantee the mapping correctness.
  */
 static void
 sock_map_release(int placement_id)
@@ -832,6 +835,8 @@ spdk_sock_write_config_json(struct spdk_json_write_ctx *w)
 			spdk_json_write_named_uint32(w, "send_buf_size", opts.send_buf_size);
 			spdk_json_write_named_bool(w, "enable_recv_pipe", opts.enable_recv_pipe);
 			spdk_json_write_named_bool(w, "enable_zerocopy_send", opts.enable_zerocopy_send);
+			spdk_json_write_named_bool(w, "enable_quickack", opts.enable_quickack);
+			spdk_json_write_named_bool(w, "enable_placement_id", opts.enable_placement_id);
 			spdk_json_write_object_end(w);
 			spdk_json_write_object_end(w);
 		} else {