+++ lib/vhost/vhost.c
@@ -888,6 +888,7 @@ static int
 vhost_parse_core_mask(const char *mask, struct spdk_cpuset *cpumask)
 {
 	int rc;
+	struct spdk_cpuset negative_vhost_mask;
 
 	if (cpumask == NULL) {
 		return -1;
@@ -904,6 +905,16 @@ vhost_parse_core_mask(const char *mask, struct spdk_cpuset *cpumask)
 		return -1;
 	}
 
+	spdk_cpuset_copy(&negative_vhost_mask, &g_vhost_core_mask);
+	spdk_cpuset_negate(&negative_vhost_mask);
+	spdk_cpuset_and(&negative_vhost_mask, cpumask);
+
+	if (spdk_cpuset_count(&negative_vhost_mask) != 0) {
+		SPDK_ERRLOG("one of selected cpu is outside of core mask(=%s)\n",
+			    spdk_cpuset_fmt(&g_vhost_core_mask));
+		return -1;
+	}
+
 	spdk_cpuset_and(cpumask, &g_vhost_core_mask);
 
 	if (spdk_cpuset_count(cpumask) == 0) {
@@ -913,26 +924,6 @@ vhost_parse_core_mask(const char *mask, struct spdk_cpuset *cpumask)
 	}
 
 	return 0;
-}
-
-static void
-vhost_setup_core_mask(void *ctx)
-{
-	struct spdk_thread *thread = spdk_get_thread();
-	spdk_cpuset_or(&g_vhost_core_mask, spdk_thread_get_cpumask(thread));
-}
-
-static void
-vhost_setup_core_mask_done(void *ctx)
-{
-	spdk_vhost_init_cb init_cb = ctx;
-
-	if (spdk_cpuset_count(&g_vhost_core_mask) == 0) {
-		init_cb(-ECHILD);
-		return;
-	}
-
-	init_cb(0);
 }
 
 static void
@@ -1562,7 +1553,8 @@ void
 spdk_vhost_init(spdk_vhost_init_cb init_cb)
 {
 	size_t len;
-	int ret;
+	uint32_t i;
+	int ret = 0;
 
 	g_vhost_init_thread = spdk_get_thread();
 	assert(g_vhost_init_thread != NULL);
@@ -1589,12 +1581,9 @@ spdk_vhost_init(spdk_vhost_init_cb init_cb)
 	}
 
 	spdk_cpuset_zero(&g_vhost_core_mask);
-
-	/* iterate threads instead of using SPDK_ENV_FOREACH_CORE to ensure that threads are really
-	 * created.
-	 */
-	spdk_for_each_thread(vhost_setup_core_mask, init_cb, vhost_setup_core_mask_done);
-	return;
+	SPDK_ENV_FOREACH_CORE(i) {
+		spdk_cpuset_set_cpu(&g_vhost_core_mask, i, true);
+	}
 out:
 	init_cb(ret);
 }