+++ module/sock/posix/posix.c
@@ -81,7 +81,7 @@ static struct spdk_sock_impl_opts g_spdk_posix_sock_impl_opts = {
 	.recv_buf_size = MIN_SO_RCVBUF_SIZE,
 	.send_buf_size = MIN_SO_SNDBUF_SIZE,
 	.enable_recv_pipe = true,
-	.enable_zerocopy_send = false,
+	.enable_zerocopy_send = true,
 	.enable_quickack = false,
 	.enable_placement_id = false,
 };
@@ -660,6 +660,7 @@ static int
 _sock_check_zcopy(struct spdk_sock *sock)
 {
 	struct spdk_posix_sock *psock = __posix_sock(sock);
+	struct spdk_posix_sock_group_impl *group = __posix_group_impl(sock->group_impl);
 	struct msghdr msgh = {};
 	uint8_t buf[sizeof(struct cmsghdr) + sizeof(struct sock_extended_err)];
 	ssize_t rc;
@@ -723,6 +724,12 @@ _sock_check_zcopy(struct spdk_sock *sock)
 				}
 			}
 
+			/* If we reaped buffer reclaim notification and sock is not in pending_recv list yet,
+			 * add it now. It allows to call socket callback and process completions */
+			if (found && !psock->pending_recv) {
+				psock->pending_recv = true;
+				TAILQ_INSERT_TAIL(&group->pending_recv, psock, link);
+			}
 		}
 	}