+++ test/scheduler/load_balancing.sh
@@ -8,26 +8,9 @@ source "$testdir/common.sh"
 
 trap 'killprocess "$spdk_pid"' EXIT
 
-export PYTHONPATH=$rootdir/test/event/scheduler
-
-declare -r scheduler=$rootdir/test/event/scheduler/scheduler
-declare -r plugin=scheduler_plugin
-
 fold_list_onto_array cpus $(parse_cpu_list <(echo "$spdk_cpus_csv"))
 # Normalize the indexes
 cpus=("${cpus[@]}")
-
-create_thread() {
-	"$rootdir/scripts/rpc.py" --plugin "$plugin" scheduler_thread_create "$@"
-}
-
-destroy_thread() {
-	"$rootdir/scripts/rpc.py" --plugin "$plugin" scheduler_thread_delete "$@"
-}
-
-active_thread() {
-	"$rootdir/scripts/rpc.py" --plugin "$plugin" scheduler_thread_set_active "$@"
-}
 
 busy() {
 	local selected_cpus cpu
@@ -124,6 +107,10 @@ balanced() {
 		else
 			active_thread "$thread0" 0
 		fi
+
+		# Give scheduler some time to spin the cpu down|up
+		sleep 0.5s
+
 		reactor_framework=$(rpc_cmd framework_get_reactors | jq -r '.reactors[]')
 		printf '* Sample %u\n' "$samples"
 		# Include main cpu to check if thread is put back on it